requirements.txt
# Langchain
langchain==0.1.12
langchain-core>=0.1.27,<0.2.0
langchain-community>=0.0.35,<0.1.0
langchain-google-genai>=0.0.11,<0.1.0
langchain-text-splitters>=0.0.2,<0.1.0

# Google Gemini
google-generativeai>=0.4.1,<0.5.0

# Vector store and data processing
chromadb>=0.4.24,<0.5.0
pandas>=2.2.0,<2.3.0
numpy>=1.26.0,<1.27.0

# Utilities
pyyaml>=6.0.1,<7.0.0
tqdm>=4.66.0,<5.0.0
python-dotenv>=1.0.0,<2.0.0
typing-extensions>=4.10.0,<5.0.0
python-dotenv>=1.0.0,<2.0.0
scikit-learn==1.4.2
tqdm>=4.66.4

# Scraping
beautifulsoup4==4.12.1
PyPDF2==3.0.1